+++ core/src/main/scala/com/qohat/brand.scala
@@ -7,10 +7,5 @@ package object brand {
     @newtype case class BrandId(value: UUID)
     @newtype case class BrandName(value: String)
     case class Brand(id: BrandId, name: BrandName)
-
-    trait Bands[F[_]] {
-        def findAll: F[List[Brand]]
-        def create(name: BrandName): F[BrandId]
-    }
 }
