+++ core/src/main/scala/com/qohat/category.scala
@@ -7,9 +7,4 @@ package object category {
     @newtype case class CategoryId(value: UUID)
     @newtype case class CategoryName(value: String)
     case class Category(id: CategoryId, name: CategoryName)
-
-    trait Categorys[F[_]] {
-        def findAll: F[List[Category]]
-        def create(name: CategoryName): F[CategoryId]
-    }
 }
