+++ domain/src/main/scala/com/qohat/domain/Brand.scala
@@ -1,0 +1,10 @@
+package com.qohat.domain
+
+import com.qohat.brand.BrandName
+import com.qohat.brand.BrandId
+import com.qohat.brand.Brand
+
+trait Bands[F[_]] {
+    def findAll: F[List[Brand]]
+    def create(name: BrandName): F[BrandId]
+}
