+++ domain/src/main/scala/com/qohat/domain/User.scala
@@ -1,0 +1,12 @@
+package com.qohat.domain
+
+trait Users[F[_]] {
+  def find(
+    username: UserName
+  ): F[Option[UserWithPassword]]
+
+  def create(
+    username: UserName,
+    password: EncryptedPassword
+  ): F[UserId]
+}
